Slight behavior change of ConfigBuilder and EnvBuilder
Change EnvBuilder to not require ExecutorService and Camino will create one
if none is available. This way there is no need to create an external
ExecutorService thus making the code cleaner in most cases.

Rename build() and expand() to buildLocal() and build() in ConfigBuilder
since expanding config should be automatic. Change CaminoApp to load from
location so it can expand config recursively.

Signed-off-by: Larry Lo <llo@turn.com>