Added support for lambda expression, member access, and custom metrics
The goal of this commit is to support custom metrics defined by user using
Camino expression language. In order to achieve that, we needed to support
user-defined functions and access to Camino classes in metric functions.
Functions are now first-class objects in Camino and are defined via lambda
expression in properties. Custom metric will specific a function to compute
metric values over PathDetail instances, meaning that access to member
fields of PathDetail, hence the support for member access operator.

Signed-off-by: Larry Lo <llo@turn.com>